Create README.md
This paper was created as part of the IB Math Analysis HL high school course and served as my Internal Assessment. The program written as part of the investigation calculates the area of a circle and an eight-petaled rhodonea curve via the Monte Carlo method. A live demo of the program is linked below.